fix: resolve bun startup module cycles
