+++ src/commands/index.ts
@@ -1,5 +1,3 @@
-import type Discord from 'discord.js';
-import { createEmbed } from '../lib/embed';
 import { cabin } from './a350x/cabin';
 import { faq } from './a350x/faq';
 import { installer } from './a350x/installer';
@@ -9,6 +7,7 @@ import { simbrief } from './a350x/simbrief';
 import { tools } from './a350x/tools';
 import { variants } from './a350x/variants';
 import { when } from './a350x/when';
+import type { CommandDefinition } from './definitions';
 import { whoosh } from './fun/whoosh';
 import { help } from './general/help';
 import { metar } from './general/metar';
@@ -28,22 +27,6 @@ import { warn } from './moderation/warn';
 import { warnings } from './moderation/warnings';
 import { whois } from './moderation/whois';
 import { msfs } from './support/msfs';
-
-export enum CommandCategories {
-    A350X = 'A350X',
-    GENERAL = 'General',
-    FUN = 'Fun',
-    MODERATION = 'Moderation',
-    SUPPORT = 'Support',
-}
-export type CommandDefinition = {
-    names: string[];
-    description: string;
-    category: CommandCategories;
-    permissions?: Discord.PermissionsString[];
-    execute: (message: Discord.Message, args: Array<string>) => Promise<any>;
-};
-export const createErrorEmbed = (description: string) => createEmbed({ color: 0xff0000, title: 'Error', description });
 
 export const commands: CommandDefinition[] = [
     when,
