+++ src/commands/moderation/dm.ts
@@ -1,5 +1,5 @@
 import { createEmbed } from '../../lib/embed';
-import { CommandCategories, type CommandDefinition, createErrorEmbed } from '../index';
+import { CommandCategories, type CommandDefinition, createErrorEmbed } from '../definitions';
 
 export const dm: CommandDefinition = {
     names: ['dm'],
