+++ src/commands/moderation/removewarning.ts
@@ -1,6 +1,6 @@
 import { createEmbed } from '../../lib/embed';
 import Warning from '../../schemas/warning';
-import { CommandCategories, type CommandDefinition, createErrorEmbed } from '../index';
+import { CommandCategories, type CommandDefinition, createErrorEmbed } from '../definitions';
 
 export const removewarning: CommandDefinition = {
     names: ['removewarning', 'rmwarn', 'deletewarning', 'delwarn'],
