+++ src/commands/moderation/rules.ts
@@ -1,7 +1,7 @@
 /* eslint-disable max-len */
 
 import { createEmbed } from '../../lib/embed';
-import { CommandCategories, type CommandDefinition } from '../index';
+import { CommandCategories, type CommandDefinition } from '../definitions';
 
 export const rules: CommandDefinition = {
     names: ['rules'],
