+++ src/constants.ts
@@ -1,0 +1,1 @@
+export const color = 0x18b1ab;
