+++ src/index.ts
@@ -16,7 +16,6 @@ const client = new Client({
     }),
 });
 
-export const color = 0x18b1ab;
 const prefix = '.';
 
 client.on('ready', (client) => {
